pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "sigma-rx7"
version = "0.1.0"
description = "Synthetic healthcare ETL/ELT pipeline for FHIR data processing"
readme = "README.md"
authors = [{name = "SigmaRx7 Team", email = "team@sigma-rx7.com"}]
license = {text = "MIT"}
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Healthcare Industry",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: Scientific/Engineering :: Medical Science Apps.",
]
requires-python = ">=3.8"
dependencies = [
    "pydantic>=2.0.0",
    "pyyaml>=6.0",
    "click>=8.0.0",
    "python-dotenv>=1.0.0",
    "sqlalchemy>=2.0.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "psycopg2-binary>=2.9.0",
    "fhir.resources>=7.0.0",
    "fhirclient>=4.0.0",
    "hl7apy>=1.3.0",
    "scikit-learn>=1.3.0",
    "joblib>=1.3.0",
    "requests>=2.31.0",
    "jsonschema>=4.0.0",
    "structlog>=23.0.0",
    "rich>=13.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
    "isort>=5.12.0",
    "factory-boy>=3.2.0",
    "faker>=18.0.0",
]

[project.scripts]
sigma-rx7 = "sigma_rx7.cli.main:main"

[project.urls]
Homepage = "https://github.com/pchj/sigma-rx7"
"Bug Tracker" = "https://github.com/pchj/sigma-rx7/issues"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
sigma_rx7 = ["config/*.yaml", "schemas/*.json"]

[tool.black]
line-length = 88
target-version = ['py38']
include = '\.pyi?$'

[tool.isort]
profile = "black"
multi_line_output = 3
line_length = 88

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "-v --tb=short"